+++ crates/crabe_math/src/shape/arc.rs
@@ -4,11 +4,11 @@ use nalgebra::Point2;
 #[derive(Debug)]
 pub struct Arc {
     /// The center point of the arc in meters.
-    pub center: Point2<f32>,
+    pub center: Point2<f64>,
     /// The radius of the arc in meters.
-    pub radius: f32,
+    pub radius: f64,
     /// The starting angle of the arc in radians.
-    pub start_angle: f32,
+    pub start_angle: f64,
     /// The ending angle of the arc in radians.
-    pub end_angle: f32,
+    pub end_angle: f64,
 }
